+++ gulpfile.js
@@ -1,7 +1,6 @@
 const watchify = require('watchify')
 const browserify = require('browserify')
 const envify = require('envify/custom')
-const disc = require('disc')
 const gulp = require('gulp')
 const source = require('vinyl-source-stream')
 const buffer = require('vinyl-buffer')
@@ -13,10 +12,7 @@ const zip = require('gulp-zip')
 const assign = require('lodash.assign')
 const livereload = require('gulp-livereload')
 const del = require('del')
-const fs = require('fs')
-const path = require('path')
 const manifest = require('./app/manifest.json')
-const mkdirp = require('mkdirp')
 const sass = require('gulp-sass')
 const autoprefixer = require('gulp-autoprefixer')
 const gulpStylelint = require('gulp-stylelint')
@@ -47,6 +43,7 @@ function gulpParallel (...args) {
 const browserPlatforms = [
   'firefox',
   'chrome',
+  'brave',
   'edge',
   'opera',
 ]
@@ -180,6 +177,7 @@ gulp.task('manifest:production', function () {
   return gulp.src([
     './dist/firefox/manifest.json',
     './dist/chrome/manifest.json',
+    './dist/brave/manifest.json',
     './dist/edge/manifest.json',
     './dist/opera/manifest.json',
   ], {base: './dist/'})
@@ -345,7 +343,7 @@ function createTasksForBuildJsExtension ({ buildJsFiles, taskPrefix, devMode, te
   const destinations = browserPlatforms.map(platform => `./dist/${platform}`)
   bundleTaskOpts = Object.assign({
     buildSourceMaps: true,
-    sourceMapDir: devMode ? './' : '../sourcemaps',
+    sourceMapDir: '../sourcemaps',
     minifyBuild: !devMode,
     buildWithFullPaths: devMode,
     watch: devMode,
@@ -374,14 +372,6 @@ function createTasksForBuildJs ({ rootDir, taskPrefix, bundleTaskOpts, destinati
 
   gulp.task(taskPrefix, gulp.series(subtasks))
 }
-
-// disc bundle analyzer tasks
-
-buildJsFiles.forEach((jsFile) => {
-  gulp.task(`disc:${jsFile}`, discTask({ label: jsFile, filename: `${jsFile}.js` }))
-})
-
-gulp.task('disc', gulp.parallel(buildJsFiles.map(jsFile => `disc:${jsFile}`)))
 
 // clean dist
 
@@ -499,6 +489,8 @@ function generateBundler (opts, performBundle) {
   }
 
   let bundler = browserify(browserifyOpts)
+    .transform('babelify')
+    .transform('brfs')
 
   if (opts.buildLib) {
     bundler = bundler.require(opts.dependenciesToBundle)
@@ -531,32 +523,6 @@ function generateBundler (opts, performBundle) {
 
   return bundler
 }
-
-function discTask (opts) {
-  opts = Object.assign({
-    buildWithFullPaths: true,
-  }, opts)
-
-  const bundler = generateBundler(opts, performBundle)
-  // output build logs to terminal
-  bundler.on('log', gutil.log)
-
-  return performBundle
-
-  function performBundle () {
-    // start "disc" build
-    const discDir = path.join(__dirname, 'disc')
-    mkdirp.sync(discDir)
-    const discPath = path.join(discDir, `${opts.label}.html`)
-
-    return (
-      bundler.bundle()
-      .pipe(disc())
-      .pipe(fs.createWriteStream(discPath))
-    )
-  }
-}
-
 
 function bundleTask (opts) {
   const bundler = generateBundler(opts, performBundle)
@@ -602,10 +568,17 @@ function bundleTask (opts) {
       }))
     }
 
-    // Finalize Source Maps (writes .map file)
+    // Finalize Source Maps
     if (opts.buildSourceMaps) {
-      buildStream = buildStream
-        .pipe(sourcemaps.write(opts.sourceMapDir))
+      if (opts.devMode) {
+        // Use inline source maps for development due to Chrome DevTools bug
+        // https://bugs.chromium.org/p/chromium/issues/detail?id=931675
+        buildStream = buildStream
+          .pipe(sourcemaps.write())
+      } else {
+        buildStream = buildStream
+          .pipe(sourcemaps.write(opts.sourceMapDir))
+      }
     }
 
     // write completed bundles