+++ ui/app/components/ui/toggle-button/index.js
@@ -1,0 +1,2 @@
+import ToggleButton from './toggle-button.component'
+module.exports = ToggleButton